+++ src/main/java/roomescape/config/WebMvcConfig.java
@@ -5,9 +5,9 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.JwtTokenProvider;
 import roomescape.controller.CheckAuthInterceptor;
 import roomescape.controller.LoginMemberArgumentResolver;
+import roomescape.util.JwtTokenProvider;
 
 @Configuration
 public class WebMvcConfig implements WebMvcConfigurer {
