+++ src/main/java/roomescape/constant/Role.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.constant;
 
 public enum Role {
     ADMIN, USER
