--- src/main/java/roomescape/dto/ReservationFilterRequest.java
@@ -1,6 +1,0 @@
-package roomescape.dto;
-
-import java.time.LocalDate;
-
-public record ReservationFilterRequest(Long memberId, Long themeId, LocalDate dateFrom, LocalDate dateTo) {
-}
