+++ src/main/java/roomescape/exception/AuthorizationException.java
@@ -1,4 +1,4 @@
-package roomescape.service;
+package roomescape.exception;
 
 public class AuthorizationException extends RuntimeException {
