+++ src/main/java/roomescape/exception/RestExceptionHandler.java
@@ -5,7 +5,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import roomescape.service.AuthorizationException;
 
 @ControllerAdvice
 public class RestExceptionHandler {
