+++ src/main/java/roomescape/repository/JdbcMemberRepository.java
@@ -7,7 +7,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.Role;
+import roomescape.constant.Role;
 import roomescape.domain.Member;
 
 @Repository
