+++ src/main/java/roomescape/util/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.util;
 
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
@@ -7,7 +7,7 @@ import java.util.Date;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import roomescape.domain.Member;
-import roomescape.service.AuthorizationException;
+import roomescape.exception.AuthorizationException;
 
 @Component
 public class JwtTokenProvider {
